+++ apps/web/src/pages/login/ui/LoginPage.tsx
@@ -1,34 +1,7 @@
-"use client";
-
-import { useState, useTransition } from "react";
-import { useRouter } from "next/navigation";
-import { Button } from "@/shared/ui/button";
-import { Input } from "@/shared/ui/input";
-import { ArrowRight, BadgeCheck, Building2, KeyRound, ShieldCheck, UserRound } from "lucide-react";
+import { BadgeCheck, Building2, ShieldCheck } from "lucide-react";
+import { LoginForm } from "./LoginForm";
 
 export function LoginPage() {
-  const router = useRouter();
-  const [isPending, startTransition] = useTransition();
-  const [employeeId, setEmployeeId] = useState("");
-  const [password, setPassword] = useState("");
-  const [errorMessage, setErrorMessage] = useState("");
-
-  const handleSubmit = (event: React.FormEvent<HTMLFormElement>) => {
-    event.preventDefault();
-
-    const isValidCredential = employeeId === "aegis-0123" && password === "021224";
-
-    if (!isValidCredential) {
-      setErrorMessage("계정 정보가 일치하지 않습니다. 인사팀에서 전달된 정보를 다시 확인하세요.");
-      return;
-    }
-
-    setErrorMessage("");
-    startTransition(() => {
-      router.push("/");
-    });
-  };
-
   return (
     <main className="relative min-h-screen overflow-hidden bg-linear-to-br from-slate-100 via-sky-50 to-cyan-100 px-4 py-10 text-slate-900 sm:px-6 lg:px-8">
       <div className="absolute inset-0 bg-[radial-gradient(circle_at_top_left,_rgba(14,165,233,0.18),_transparent_32%),radial-gradient(circle_at_bottom_right,_rgba(16,185,129,0.15),_transparent_28%)]" />
@@ -113,82 +86,7 @@ export function LoginPage() {
             </p>
           </div>
 
-          <form className="mt-8 space-y-5" onSubmit={handleSubmit}>
-            <div className="space-y-2">
-              <label htmlFor="employee-id" className="text-sm font-medium text-slate-700">
-                아이디
-              </label>
-              <div className="relative">
-                <UserRound className="pointer-events-none absolute top-1/2 left-3 size-4 -translate-y-1/2 text-slate-400" />
-                <Input
-                  id="employee-id"
-                  type="text"
-                  placeholder="aegis-0123"
-                  className="pl-10"
-                  autoComplete="username"
-                  value={employeeId}
-                  onChange={(event) => setEmployeeId(event.target.value)}
-                />
-              </div>
-            </div>
-
-            <div className="space-y-2">
-              <div className="flex items-center justify-between">
-                <label htmlFor="employee-password" className="text-sm font-medium text-slate-700">
-                  비밀번호
-                </label>
-                <button
-                  type="button"
-                  className="text-xs font-medium text-sky-700 transition hover:text-sky-800"
-                >
-                  비밀번호 재설정
-                </button>
-              </div>
-              <div className="relative">
-                <KeyRound className="pointer-events-none absolute top-1/2 left-3 size-4 -translate-y-1/2 text-slate-400" />
-                <Input
-                  id="employee-password"
-                  type="password"
-                  placeholder="비밀번호를 입력하세요"
-                  className="pl-10"
-                  autoComplete="current-password"
-                  value={password}
-                  onChange={(event) => setPassword(event.target.value)}
-                />
-              </div>
-            </div>
-
-            <div className="rounded-2xl border border-emerald-100 bg-emerald-50 px-4 py-3 text-sm text-emerald-900">
-              보안 상태 정상. 현재 네트워크는 사내망으로 인식되었습니다.
-            </div>
-
-            {errorMessage ? (
-              <div className="rounded-2xl border border-rose-100 bg-rose-50 px-4 py-3 text-sm text-rose-900">
-                {errorMessage}
-              </div>
-            ) : null}
-
-            <div className="space-y-3 pt-1">
-              <Button
-                type="submit"
-                size="lg"
-                disabled={isPending}
-                className="h-11 w-full bg-sky-600 font-semibold text-white shadow-lg shadow-sky-200 hover:bg-sky-500"
-              >
-                {isPending ? "이동 중..." : "로그인"}
-                <ArrowRight className="size-4 transition-transform group-hover/button:translate-x-0.5" />
-              </Button>
-
-              <Button
-                type="button"
-                variant="outline"
-                size="lg"
-                className="h-11 w-full border-slate-200 bg-white text-slate-700 hover:bg-slate-50"
-              >
-                게스트 로그인
-              </Button>
-            </div>
-          </form>
+          <LoginForm />
 
           <div className="mt-8 flex items-center justify-between border-t border-slate-200 pt-5 text-xs text-slate-500">
             <span>System Status: Online</span>
